feat: nav-charge mechanic, Mühle game, UI consistency pass
Nav-charge navigation replaces instant cursor jumps with a hold-to-charge
mechanic: target field shows an SVG border that draws progressively around
the rectangle (stroke-dasharray) until it closes, giving tetraplegic patients
visual prediction of when the next jump happens. Joystick intensity (analog
0.0-1.0) controls charge speed; keyboard has dual mode (Direkt for caretaker,
Joystick-Sim via J key for remote testing). Default charge duration 1000ms,
configurable via settings.

Adds Mühle (Nine Men's Morris) as 6th game with 3 AI difficulty levels
(minimax + alpha-beta), multiplayer toggle, and spatial navigation mapping
arrow keys to geometric position neighbors on the board.

Firmware hardening: strict linting (clang-tidy + cppcheck, fail-on-defect=low),
null-pointer guards in sensor task, OTA fresh-flash detection via LittleFS
marker file (prevents stale version tracking after USB flash), intensity
field in nav WebSocket messages with nav_hold/nav_release event types.

UI consistency sweep across all 6 games: unified icon-based footer (↻ ↩ 💡
⌨ 👥 📺 🏠), consistent footer color (black static instead of mixed blue/
black), charge navigation in all games (previously only solitaire + mühle),
multiplayer mode in chess/vier-gewinnt/mühle, larger Mühle pieces with
always-visible cursor rings in dedicated highlight layer.

+++ firmware/arduino/src/wifi_manager.cpp
@@ -187,7 +187,7 @@ std::pair<int, String> WiFiManager::getRSSI() {
         return {0, ""};
     }
 
-    int rssi = WiFi.RSSI();
+    int rssi = static_cast<int>(WiFi.RSSI());  // NOLINT(bugprone-signed-char-misuse) RSSI is signed
     String label;
     if      (rssi >= -50) label = "Perfekt";
     else if (rssi >= -60) label = "Gut";
